+++ .changeset/tangy-deer-fetch.md
@@ -4,3 +4,4 @@
 
 - feat(cli): remove `--no-preconditions` option from `sv add`
 - feat(cli): add `--no-git-check` option to `sv add`. With this flag, even if some files are dirty, no prompt will be shown
+- feat(cli): type `Precondition` was replaced with `Verification`
